+++ examples/sdl_sample/src/sdl_sample.cpp
@@ -50,7 +50,7 @@ class SDLSample : public std::enable_shared_from_this<SDLSample>,
       cam_config.fps = 30;
       auto video_source = sora::CreateCameraDeviceCapturer(cam_config);
 
-      std::string video_track_id = rtc::CreateRandomString(16);
+      std::string video_track_id = webrtc::CreateRandomString(16);
       video_track_ = context_->peer_connection_factory()->CreateVideoTrack(
           video_source, video_track_id);
       if (config_.show_me) {
@@ -58,7 +58,7 @@ class SDLSample : public std::enable_shared_from_this<SDLSample>,
       }
     }
     if (config_.audio && config_.role != "recvonly") {
-      std::string audio_track_id = rtc::CreateRandomString(16);
+      std::string audio_track_id = webrtc::CreateRandomString(16);
       audio_track_ = context_->peer_connection_factory()->CreateAudioTrack(
           audio_track_id, context_->peer_connection_factory()
                               ->CreateAudioSource(webrtc::AudioOptions())
@@ -101,7 +101,7 @@ class SDLSample : public std::enable_shared_from_this<SDLSample>,
   }
 
   void OnSetOffer(std::string offer) override {
-    std::string stream_id = rtc::CreateRandomString(16);
+    std::string stream_id = webrtc::CreateRandomString(16);
     if (audio_track_ != nullptr) {
       webrtc::RTCErrorOr<webrtc::scoped_refptr<webrtc::RtpSenderInterface>>
           audio_result =
@@ -199,7 +199,7 @@ int main(int argc, char* argv[]) {
   app.set_help_all_flag("--help-all",
                         "Print help message for all modes and exit");
 
-  int log_level = (int)rtc::LS_ERROR;
+  int log_level = (int)webrtc::LS_ERROR;
   auto log_level_map = std::vector<std::pair<std::string, int>>(
       {{"verbose", 0}, {"info", 1}, {"warning", 2}, {"error", 3}, {"none", 4}});
   app.add_option("--log-level", log_level, "Log severity level threshold")
@@ -244,10 +244,10 @@ int main(int argc, char* argv[]) {
     config.metadata = boost::json::parse(metadata);
   }
 
-  if (log_level != rtc::LS_NONE) {
-    rtc::LogMessage::LogToDebug((rtc::LoggingSeverity)log_level);
-    rtc::LogMessage::LogTimestamps();
-    rtc::LogMessage::LogThreads();
+  if (log_level != webrtc::LS_NONE) {
+    webrtc::LogMessage::LogToDebug((webrtc::LoggingSeverity)log_level);
+    webrtc::LogMessage::LogTimestamps();
+    webrtc::LogMessage::LogThreads();
   }
 
   auto context =